Changes associated with request from Shaun regarding pull request Now done and compiles, Not tested. Need to merge changes with another branch to conduct testing.
