+++ industrial_extrinsic_cal/include/industrial_extrinsic_cal/ceres_blocks.h
@@ -201,6 +201,23 @@ public:
 
 };//end class
 
+ // dangling debugging functions
+/*! @brief displays a pose from the extrinsic parameters 
+ *   @param extrinsics an array of parameters representing the 6D pose
+ *   @param message a string that describes the pose to be shown 
+ */
+ void  showPose(P_BLOCK extrinsics, std::string message);
+/*! @brief displays a pose 
+ *   @param pose the 6D pose to be shown
+ *   @param message a string that describes the pose to be shown 
+ */
+ void  showPose(Pose6d pose, std::string message);
+/*! @brief displays the intrinsic parameters
+ *   @param intrinsics an array of parameters representing the focal length, optical center, and distortion parameters
+ *   @param num_param, sometimes there is no distortion, so this is either 4 or 9 depending
+ */
+ void  showIntrinsics(P_BLOCK intrinsics, int num_param);
+
 }// end namespace industrial_extrinsic_cal
 
 #endif /* CERES_BLOCKS_H_ */
