+++ industrial_extrinsic_cal/include/industrial_extrinsic_cal/ceres_costs_utils.h
@@ -19,6 +19,7 @@
 #ifndef CERES_COSTS_UTILS_H_
 #define CERES_COSTS_UTILS_H_
 
+#include <string>
 namespace industrial_extrinsic_cal
 {
   // create an enumeration of all the cost functions
@@ -45,61 +46,24 @@ namespace industrial_extrinsic_cal
       LinkCircleTargetCameraReprjError,
       LinkCircleTargetCameraReprjErrorPK,
       LinkCameraCircleTargetReprjError,
-      LinkCameraCircleTargetReprjErrorPK
+      LinkCameraCircleTargetReprjErrorPK,
+      NullCostType
     };
   }// end of namespace cost_functions
   typedef cost_functions::Cost_function Cost_function;
 
-  Cost_function string2CostType(std::string &cost_type_str)
-  {
-    if(cost_type_str == "CameraReprjErrorWithDistortion") return(cost_functions::CameraReprjErrorWithDistortion);
-    if(cost_type_str == "CameraReprjErrorWithDistortionPK") return(cost_functions::CameraReprjErrorWithDistortionPK);
-    if(cost_type_str == "CameraReprjError") return(cost_functions::CameraReprjError);
-    if(cost_type_str == "CameraReprjErrorPK") return(cost_functions::CameraReprjErrorPK);
-    if(cost_type_str == "TargetCameraReprjError") return(cost_functions::TargetCameraReprjError);
-    if(cost_type_str == "TargetCameraReprjErrorPK") return(cost_functions::TargetCameraReprjErrorPK);
-    if(cost_type_str == "LinkTargetCameraReprjError") return(cost_functions::LinkTargetCameraReprjError);
-    if(cost_type_str == "LinkTargetCameraReprjErrorPK") return(cost_functions::LinkTargetCameraReprjErrorPK);
-    if(cost_type_str == "LinkCameraTargetReprjError") return(cost_functions::LinkCameraTargetReprjError);
-    if(cost_type_str == "LinkCameraTargetReprjErrorPK") return(cost_functions::LinkCameraTargetReprjErrorPK);
-    if(cost_type_str == "CircleCameraReprjErrorWithDistortion") return(cost_functions::CircleCameraReprjErrorWithDistortion);
-    if(cost_type_str == "CircleCameraReprjErrorWithDistortionPK") return(cost_functions::CircleCameraReprjErrorWithDistortionPK);
-    if(cost_type_str == "CircleCameraReprjError") return(cost_functions::CircleCameraReprjError);
-    if(cost_type_str == "CircleCameraReprjErrorPK") return(cost_functions::CircleCameraReprjErrorPK);
-    if(cost_type_str == "CircleTargetCameraReprjErrorWithDistortion") return(cost_functions::CircleTargetCameraReprjErrorWithDistortion);
-    if(cost_type_str == "CircleTargetCameraReprjErrorWithDistortionPK") return(cost_functions::CircleTargetCameraReprjErrorWithDistortionPK);
-    if(cost_type_str == "CircleTargetCameraReprjError") return(cost_functions::CircleTargetCameraReprjError);
-    if(cost_type_str == "CircleTargetCameraReprjErrorPK") return(cost_functions::CircleTargetCameraReprjErrorPK);
-    if(cost_type_str == "LinkCircleTargetCameraReprjError") return(cost_functions::LinkCircleTargetCameraReprjError);
-    if(cost_type_str == "LinkCircleTargetCameraReprjErrorPK") return(cost_functions::LinkCircleTargetCameraReprjErrorPK);
-    if(cost_type_str == "LinkCameraCircleTargetReprjError") return(cost_functions::LinkCameraCircleTargetReprjError);
-    if(cost_type_str == "LinkCameraCircleTargetReprjErrorPK") return(cost_functions::LinkCameraCircleTargetReprjErrorPK);
-  }
+  // prototypes of functions 
 
-  std::string costType2String(Cost_function cost_type)
-  {
-    if(cost_type == cost_functions::CameraReprjErrorWithDistortion) return("CameraReprjErrorWithDistortion");
-    if(cost_type == cost_functions::CameraReprjErrorWithDistortionPK) return("CameraReprjErrorWithDistortionPK");
-    if(cost_type == cost_functions::CameraReprjError) return("CameraReprjError");
-    if(cost_type == cost_functions::CameraReprjErrorPK) return("CameraReprjErrorPK");
-    if(cost_type == cost_functions::TargetCameraReprjError) return("TargetCameraReprjError");
-    if(cost_type == cost_functions::TargetCameraReprjErrorPK) return("TargetCameraReprjErrorPK");
-    if(cost_type == cost_functions::LinkTargetCameraReprjError) return("LinkTargetCameraReprjError");
-    if(cost_type == cost_functions::LinkTargetCameraReprjErrorPK) return("LinkTargetCameraReprjErrorPK");
-    if(cost_type == cost_functions::LinkCameraTargetReprjError) return("LinkCameraTargetReprjError");
-    if(cost_type == cost_functions::LinkCameraTargetReprjErrorPK) return("LinkCameraTargetReprjErrorPK");
-    if(cost_type == cost_functions::CircleCameraReprjErrorWithDistortion) return("CircleCameraReprjErrorWithDistortion");
-    if(cost_type == cost_functions::CircleCameraReprjErrorWithDistortionPK) return("CircleCameraReprjErrorWithDistortionPK");
-    if(cost_type == cost_functions::CircleCameraReprjError) return("CircleCameraReprjError");
-    if(cost_type == cost_functions::CircleCameraReprjErrorPK) return("CircleCameraReprjErrorPK");
-    if(cost_type == cost_functions::CircleTargetCameraReprjErrorWithDistortion) return("CircleTargetCameraReprjErrorWithDistortion");
-    if(cost_type == cost_functions::CircleTargetCameraReprjErrorWithDistortionPK) return("CircleTargetCameraReprjErrorWithDistortionPK");
-    if(cost_type == cost_functions::CircleTargetCameraReprjError) return("CircleTargetCameraReprjError");
-    if(cost_type == cost_functions::CircleTargetCameraReprjErrorPK) return("CircleTargetCameraReprjErrorPK");
-    if(cost_type == cost_functions::LinkCircleTargetCameraReprjError) return("LinkCircleTargetCameraReprjError");
-    if(cost_type == cost_functions::LinkCircleTargetCameraReprjErrorPK) return("LinkCircleTargetCameraReprjErrorPK");
-    if(cost_type == cost_functions::LinkCameraCircleTargetReprjError) return("LinkCameraCircleTargetReprjError");
-    if(cost_type == cost_functions::LinkCameraCircleTargetReprjErrorPK) return("LinkCameraCircleTargetReprjErrorPK");
-  }
+  /*! @brief converts a string to a cost type 
+   *   @param cost_type_str The cost type string 
+   *   @returns The cost function type from the enumeration
+   */
+  Cost_function string2CostType(std::string &cost_type_str);
+  /*! @brief converts a cost type to a string
+   *   @param cost_type The cost type 
+   *   @returns The cost function type as a string
+   */
+  std::string costType2String(Cost_function cost_type);
+
 } // end of namespace industrial_extrinsic_cal
 #endif
