+++ industrial_extrinsic_cal/src/ceres_blocks.cpp
@@ -32,14 +32,39 @@ namespace industrial_extrinsic_cal
     pose.getEulerZYX(ez_yaw,ey_pitch,ex_roll);
     pose.getQuaternion(qx, qy, qz, qw);
     ROS_INFO("%s =[\n %6.3lf  %6.3lf  %6.3lf  %6.3lf\n  %6.3lf  %6.3lf  %6.3lf  %6.3lf\n  %6.3lf  %6.3lf %6.3lf  %6.3lf\n  %6.3lf  %6.3lf %6.3lf  %6.3lf];\n rpy= %lf %lf %lf\n quat= %6.3lf %6.3lf %6.3lf %6.3lf ",
-	      message.c_str(),
-	      basis[0][0],basis[0][1], basis[0][2], pose.x,
+	     message.c_str(),
+	     basis[0][0],basis[0][1], basis[0][2], pose.x,
 	     basis[1][0],basis[1][1], basis[1][2], pose.y,
-	      basis[2][0],basis[2][1], basis[2][2], pose.z,
+	     basis[2][0],basis[2][1], basis[2][2], pose.z,
 	     0.0, 0.0, 0.0, 1.0,
-	      ez_yaw, ey_pitch, ex_roll,
-	      qx,qy,qz,qw);
+	     ez_yaw, ey_pitch, ex_roll,
+	     qx,qy,qz,qw);
   }
+  
+  void  showPose(P_BLOCK extrinsics, std::string message){
+    double ax,ay,az,px,py,pz;
+    ax  = extrinsics[0]; 
+    ay  = extrinsics[1]; 
+    az  = extrinsics[2]; 
+    px  = extrinsics[3]; 
+    py  = extrinsics[4]; 
+    pz  = extrinsics[5];
+    Pose6d pose(px,py,pz,ax,ay,az);
+    tf::Matrix3x3 basis = pose.getBasis();
+    double ez_yaw, ey_pitch, ex_roll;
+    double qx, qy, qz, qw;
+    pose.getEulerZYX(ez_yaw,ey_pitch,ex_roll);
+    pose.getQuaternion(qx, qy, qz, qw);
+    ROS_INFO("%s =[\n %6.3lf  %6.3lf  %6.3lf  %6.3lf\n  %6.3lf  %6.3lf  %6.3lf  %6.3lf\n  %6.3lf  %6.3lf %6.3lf  %6.3lf\n  %6.3lf  %6.3lf %6.3lf  %6.3lf];\n rpy= %6.3lf %6.3lf %6.3lf\n quat= %6.3lf  %6.3lf  %6.3lf %6.3lf ",
+	     message.c_str(),
+	     basis[0][0],basis[0][1], basis[0][2],px,
+	     basis[1][0],basis[1][1], basis[1][2],py,
+	     basis[2][0],basis[2][1], basis[2][2],pz,
+	     0.0, 0.0, 0.0, 1.0,
+	     ez_yaw, ey_pitch, ex_roll,
+	     qx, qy, qz, qw);
+  }
+
   void  showIntrinsics(P_BLOCK intrinsics, int num_param){
     double fx,fy,cx,cy,k1,k2,k3,p1,p2;
     fx  = intrinsics[0]; /** focal length x */
